requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Django==4.1.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
google-api-core==2.10.2
google-api-python-client==2.48.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
Google-Images-Search==1.4.6
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.4
numpy==1.23.4
openai==0.23.1
openpyxl==3.0.10
pandas==1.5.0
pandas-stubs==1.5.0.221012
Pillow==9.2.0
protobuf==4.21.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyfiglet==0.8.post1
pyparsing==3.0.9
python-dateutil==2.8.2
python-resize-image==1.1.20
pytz==2022.4
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
termcolor==1.1.0
tqdm==4.64.1
types-pytz==2022.4.0.0
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
